requirements.txt
torch==2.0.1
torchvision==0.15.2
librosa==0.10.1
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
scikit-learn==1.3.2
tqdm==4.66.1
matplotlib==3.7.2
seaborn==0.12.2
